Restructure display/log and add some comments
