+++ client/display/log/mod.rs
@@ -1,18 +1,21 @@
 use std::collections::BTreeMap;
-use std::fs::File;
-use std::io::{self, Stdout};
 
-use anyhow::Result;
 use comfy_table::*;
-use snap::read::FrameDecoder;
 
-use pueue_lib::log::{get_log_file_handles, read_last_lines};
 use pueue_lib::network::message::TaskLogMessage;
 use pueue_lib::settings::Settings;
 use pueue_lib::task::{Task, TaskResult, TaskStatus};
 
-use super::{colors::Colors, helper::*};
+use super::colors::Colors;
 use crate::cli::SubCommand;
+
+mod json;
+mod local;
+mod remote;
+
+use json::*;
+use local::*;
+use remote::*;
 
 /// Print the log ouput of finished tasks.
 /// Either print the logs of every task
@@ -57,14 +60,14 @@ pub fn print_logs(
         return;
     }
 
-    // Determine, whether we should draw everything or only a part of the log output.
-    // None implicates that all lines are printed
+    // Determine, whether we should print everything or only a part of the log output.
+    // `None` implicates that all lines are printed.
     let lines = if full {
         None
     } else if let Some(lines) = lines {
         Some(lines)
     } else {
-        // By default only some lines are shown per task, if multiple tasks exist.
+        // By default, only some lines are shown per task if multiple tasks exist.
         // For a single task, the whole log output is shown.
         if task_logs.len() > 1 {
             Some(15)
@@ -73,7 +76,7 @@ pub fn print_logs(
         }
     };
 
-    // Do the actual log printing
+    // Iterate over each task and print the respective log.
     let mut task_iter = task_logs.iter_mut().peekable();
     while let Some((_, mut task_log)) = task_iter.next() {
         print_log(&mut task_log, colors, settings, lines);
@@ -183,106 +186,4 @@ fn print_task_info(task: &Task, colors: &Colors) {
     first_column.set_padding((0, 0));
 
     println!("{}", table);
-}
-
-/// The daemon didn't send any log output, thereby we didn't request any.
-/// If that's the case, read the log files from the local pueue directory
-fn print_local_log(task_id: usize, colors: &Colors, settings: &Settings, lines: Option<usize>) {
-    let (mut stdout_file, mut stderr_file) =
-        match get_log_file_handles(task_id, &settings.shared.pueue_directory()) {
-            Ok((stdout, stderr)) => (stdout, stderr),
-            Err(err) => {
-                println!("Failed to get log file handles: {}", err);
-                return;
-            }
-        };
-    // Stdout handler to directly write log file output to io::stdout
-    // without having to load anything into memory.
-    let mut stdout = io::stdout();
-
-    print_local_file(
-        &mut stdout,
-        &mut stdout_file,
-        &lines,
-        style_text("stdout:", Some(colors.green()), Some(Attribute::Bold)),
-    );
-
-    print_local_file(
-        &mut stdout,
-        &mut stderr_file,
-        &lines,
-        style_text("stderr:", Some(colors.red()), Some(Attribute::Bold)),
-    );
-}
-
-/// Print a local log file.
-/// This is usually either the stdout or the stderr
-pub fn print_local_file(stdout: &mut Stdout, file: &mut File, lines: &Option<usize>, text: String) {
-    if let Ok(metadata) = file.metadata() {
-        if metadata.len() != 0 {
-            // Don't print a newline between the task information and the first output
-            println!("\n{}", text);
-
-            // Only print the last lines if requested
-            if let Some(lines) = lines {
-                println!("{}", read_last_lines(file, *lines));
-                return;
-            }
-
-            // Print everything
-            if let Err(err) = io::copy(file, stdout) {
-                println!("Failed reading local log file: {}", err);
-            };
-        }
-    }
-}
-
-/// Prints log output received from the daemon.
-/// We can safely call .unwrap() on stdout and stderr in here, since this
-/// branch is always called after ensuring that both are `Some`.
-pub fn print_remote_log(task_log: &TaskLogMessage, colors: &Colors) {
-    // Save whether stdout was printed, so we can add a newline between outputs.
-    if let Some(bytes) = task_log.stdout.as_ref() {
-        if !bytes.is_empty() {
-            println!(
-                "\n{}",
-                style_text("stdout: ", Some(colors.green()), Some(Attribute::Bold))
-            );
-
-            if let Err(err) = decompress_and_print_remote_log(bytes) {
-                println!("Error while parsing stdout: {}", err);
-            }
-        }
-    }
-
-    if let Some(bytes) = task_log.stderr.as_ref() {
-        if !bytes.is_empty() {
-            println!(
-                "\n{}",
-                style_text("stderr: ", Some(colors.red()), Some(Attribute::Bold))
-            );
-
-            if let Err(err) = decompress_and_print_remote_log(bytes) {
-                println!("Error while parsing stderr: {}", err);
-            };
-        }
-    }
-}
-
-/// We cannot easily stream log output from the client to the daemon (yet).
-/// Right now, stdout and stderr are compressed in the daemon and sent as a single payload to the
-/// client. In here, we take that payload, decompress it and stream it it directly to stdout.
-fn decompress_and_print_remote_log(bytes: &[u8]) -> Result<()> {
-    let mut decompressor = FrameDecoder::new(bytes);
-
-    let stdout = io::stdout();
-    let mut write = stdout.lock();
-    io::copy(&mut decompressor, &mut write)?;
-
-    Ok(())
-}
-
-fn print_log_json(task_logs: BTreeMap<usize, TaskLogMessage>, settings: &Settings) {
-    println!("{}", serde_json::to_string(&task_logs).unwrap());
-    return;
 }
